+++ docs/source/changelog.rst
@@ -9,6 +9,7 @@ Changelog
 Version (dev)
 -------------
 * [HammerAddons#237](https://github.com/TeamSpen210/HammerAddons/issues/237): FGD model helpers should override each other.
+* Fix #20: VTF.compute_mipmaps() not working for cubemaps.
 
 --------------
 Version 2.3.14